correction_of_comments
